refactor(agentic-ai): extract Foundry adapter request/response mappers
AnthropicOnFoundryChatModel had grown to ~400 lines mixing three concerns:
orchestration / exception translation, langchain4j → Anthropic request
translation, and Anthropic → langchain4j response translation. Splits the
two translation halves into focused, package-private collaborator classes:

  - AnthropicOnFoundryRequestMapper: ChatRequest → MessageCreateParams
    (system extraction, tool-result grouping, tool-use block construction,
    parameter resolution including the 1024 maxTokens fallback and
    ChatRequest-overrides-config rule)
  - AnthropicOnFoundryResponseMapper: Message → ChatResponse
    (text aggregation, tool-use → ToolExecutionRequest, StopReason →
    FinishReason, Usage → TokenUsage)

The adapter shrinks to its true responsibility: orchestrate the chat call
and translate the seven AnthropicException subtypes (plus AnthropicIoException
→ TRANSPORT_ERROR and the AnthropicException fallback → ANTHROPIC_ERROR).

Per-translation tests move to focused test classes
(AnthropicOnFoundryRequestMapperTest, AnthropicOnFoundryResponseMapperTest),
calling the mappers directly instead of going through Mockito ArgumentCaptor
on the AnthropicClient. The adapter test retains the eight exception-
translation tests + a smoke test.

No behavior change.
