+++ connectors/agentic-ai/src/test/java/io/camunda/connector/agenticai/azurefoundry/langchain4j/AnthropicOnFoundryChatModelTest.java
@@ -9,7 +9,6 @@ package io.camunda.connector.agenticai.azurefoundry.langchain4j;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.anthropic.client.AnthropicClient;
@@ -24,17 +23,11 @@ import com.anthropic.errors.RateLimitException;
 import com.anthropic.errors.UnauthorizedException;
 import com.anthropic.errors.UnprocessableEntityException;
 import com.anthropic.models.messages.ContentBlock;
-import com.anthropic.models.messages.ContentBlockParam;
-import com.anthropic.models.messages.DirectCaller;
 import com.anthropic.models.messages.Message;
 import com.anthropic.models.messages.MessageCreateParams;
-import com.anthropic.models.messages.MessageParam;
 import com.anthropic.models.messages.Model;
 import com.anthropic.models.messages.StopReason;
 import com.anthropic.models.messages.TextBlock;
-import com.anthropic.models.messages.ToolResultBlockParam;
-import com.anthropic.models.messages.ToolUseBlock;
-import com.anthropic.models.messages.ToolUseBlockParam;
 import com.anthropic.models.messages.Usage;
 import com.anthropic.services.blocking.MessageService;
 import dev.langchain4j.agent.tool.ToolExecutionRequest;
@@ -46,7 +39,6 @@ import dev.langchain4j.data.message.UserMessage;
 import dev.langchain4j.model.chat.request.ChatRequest;
 import dev.langchain4j.model.chat.response.ChatResponse;
 import dev.langchain4j.model.output.FinishReason;
-import dev.langchain4j.model.output.TokenUsage;
 import io.camunda.connector.agenticai.aiagent.model.request.provider.AzureFoundryProviderConfiguration.AzureAiFoundryModel.AnthropicModel;
 import io.camunda.connector.agenticai.aiagent.model.request.provider.AzureFoundryProviderConfiguration.AzureAiFoundryModel.AnthropicModel.AnthropicModelParameters;
 import io.camunda.connector.api.error.ConnectorException;
@@ -56,8 +48,6 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -68,8 +58,6 @@ class AnthropicOnFoundryChatModelTest {
 
   @Mock private AnthropicClient mockClient;
   @Mock private MessageService mockMessageService;
-
-  @Captor private ArgumentCaptor<MessageCreateParams> paramsCaptor;
 
   private AnthropicOnFoundryChatModel adapter;
 
@@ -106,16 +94,6 @@ class AnthropicOnFoundryChatModelTest {
     return TextBlock.builder().text(text).citations(Optional.empty()).build();
   }
 
-  /** Build a minimal ToolUseBlock with all required fields set. */
-  private static ToolUseBlock toolUseBlock(String id, String name, JsonValue input) {
-    return ToolUseBlock.builder()
-        .id(id)
-        .name(name)
-        .input(input)
-        .caller(ToolUseBlock.Caller.ofDirect(DirectCaller.builder().build()))
-        .build();
-  }
-
   /** Build a Message with all required fields set. */
   private static Message message(
       String id, List<ContentBlock> blocks, StopReason stopReason, Usage usageValue) {
@@ -134,77 +112,8 @@ class AnthropicOnFoundryChatModelTest {
   }
 
   // -------------------------------------------------------------------------
-  // Tests
+  // Smoke test — end-to-end through adapter (request + response path)
   // -------------------------------------------------------------------------
-
-  @Test
-  void translates_text_response_to_ai_message() {
-    // given
-    var contentBlock = ContentBlock.ofText(textBlock("Hello, world!"));
-    var message = message("msg_001", List.of(contentBlock), StopReason.END_TURN, usage(10, 5));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request = ChatRequest.builder().messages(List.of(new UserMessage("Hi"))).build();
-
-    // when
-    ChatResponse response = adapter.chat(request);
-
-    // then
-    assertThat(response.aiMessage().text()).isEqualTo("Hello, world!");
-    assertThat(response.finishReason()).isEqualTo(FinishReason.STOP);
-    TokenUsage tokenUsage = response.tokenUsage();
-    assertThat(tokenUsage.inputTokenCount()).isEqualTo(10);
-    assertThat(tokenUsage.outputTokenCount()).isEqualTo(5);
-    assertThat(tokenUsage.totalTokenCount()).isEqualTo(15);
-  }
-
-  @Test
-  void translates_tool_use_response_to_tool_execution_requests() {
-    // given
-    var toolUseBlock =
-        toolUseBlock(
-            "toolu_01XFDUDYJgAACTU9bRTQE47m",
-            "get_weather",
-            JsonValue.from(java.util.Map.of("location", "San Francisco, CA")));
-    var contentBlock = ContentBlock.ofToolUse(toolUseBlock);
-    var message = message("msg_002", List.of(contentBlock), StopReason.TOOL_USE, usage(20, 8));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder().messages(List.of(new UserMessage("What is the weather?"))).build();
-
-    // when
-    ChatResponse response = adapter.chat(request);
-
-    // then
-    assertThat(response.finishReason()).isEqualTo(FinishReason.TOOL_EXECUTION);
-    AiMessage aiMessage = response.aiMessage();
-    assertThat(aiMessage.hasToolExecutionRequests()).isTrue();
-    List<ToolExecutionRequest> toolRequests = aiMessage.toolExecutionRequests();
-    assertThat(toolRequests).hasSize(1);
-    ToolExecutionRequest toolRequest = toolRequests.get(0);
-    assertThat(toolRequest.id()).isEqualTo("toolu_01XFDUDYJgAACTU9bRTQE47m");
-    assertThat(toolRequest.name()).isEqualTo("get_weather");
-    assertThat(toolRequest.arguments()).contains("San Francisco, CA");
-  }
-
-  @Test
-  void translates_max_tokens_stop_reason_to_LENGTH() {
-    // given
-    var contentBlock = ContentBlock.ofText(textBlock("Partial response..."));
-    var message = message("msg_003", List.of(contentBlock), StopReason.MAX_TOKENS, usage(15, 1024));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder().messages(List.of(new UserMessage("Tell me everything"))).build();
-
-    // when
-    ChatResponse response = adapter.chat(request);
-
-    // then
-    assertThat(response.finishReason()).isEqualTo(FinishReason.LENGTH);
-    assertThat(response.aiMessage().text()).isEqualTo("Partial response...");
-  }
 
   @Test
   void translates_system_and_tool_result_messages_in_request() {
@@ -241,6 +150,10 @@ class AnthropicOnFoundryChatModelTest {
     assertThat(response.aiMessage().text()).isEqualTo("Done");
   }
 
+  // -------------------------------------------------------------------------
+  // Exception translation tests
+  // -------------------------------------------------------------------------
+
   @Test
   void bad_request_becomes_connector_input_exception() {
     // given
@@ -268,41 +181,6 @@ class AnthropicOnFoundryChatModelTest {
     // when / then
     assertThatThrownBy(() -> adapter.chat(request))
         .isInstanceOf(ConnectorInputException.class)
-        .hasCause(ex);
-  }
-
-  @Test
-  void rate_limit_becomes_retryable_connector_exception() {
-    // given
-    var ex = RateLimitException.builder().headers(EMPTY_HEADERS).body(JsonValue.from(null)).build();
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenThrow(ex);
-
-    var request = ChatRequest.builder().messages(List.of(new UserMessage("hello"))).build();
-
-    // when / then
-    assertThatThrownBy(() -> adapter.chat(request))
-        .isInstanceOf(ConnectorException.class)
-        .isNotInstanceOf(ConnectorInputException.class)
-        .hasCause(ex);
-  }
-
-  @Test
-  void internal_server_error_becomes_retryable_connector_exception() {
-    // given
-    var ex =
-        InternalServerException.builder()
-            .statusCode(500)
-            .headers(EMPTY_HEADERS)
-            .body(JsonValue.from(null))
-            .build();
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenThrow(ex);
-
-    var request = ChatRequest.builder().messages(List.of(new UserMessage("hello"))).build();
-
-    // when / then
-    assertThatThrownBy(() -> adapter.chat(request))
-        .isInstanceOf(ConnectorException.class)
-        .isNotInstanceOf(ConnectorInputException.class)
         .hasCause(ex);
   }
 
@@ -357,6 +235,41 @@ class AnthropicOnFoundryChatModelTest {
   }
 
   @Test
+  void rate_limit_becomes_retryable_connector_exception() {
+    // given
+    var ex = RateLimitException.builder().headers(EMPTY_HEADERS).body(JsonValue.from(null)).build();
+    when(mockMessageService.create(any(MessageCreateParams.class))).thenThrow(ex);
+
+    var request = ChatRequest.builder().messages(List.of(new UserMessage("hello"))).build();
+
+    // when / then
+    assertThatThrownBy(() -> adapter.chat(request))
+        .isInstanceOf(ConnectorException.class)
+        .isNotInstanceOf(ConnectorInputException.class)
+        .hasCause(ex);
+  }
+
+  @Test
+  void internal_server_error_becomes_retryable_connector_exception() {
+    // given
+    var ex =
+        InternalServerException.builder()
+            .statusCode(500)
+            .headers(EMPTY_HEADERS)
+            .body(JsonValue.from(null))
+            .build();
+    when(mockMessageService.create(any(MessageCreateParams.class))).thenThrow(ex);
+
+    var request = ChatRequest.builder().messages(List.of(new UserMessage("hello"))).build();
+
+    // when / then
+    assertThatThrownBy(() -> adapter.chat(request))
+        .isInstanceOf(ConnectorException.class)
+        .isNotInstanceOf(ConnectorInputException.class)
+        .hasCause(ex);
+  }
+
+  @Test
   void transport_io_failure_becomes_retryable_connector_exception() {
     // given
     var cause = new java.io.IOException("connect refused");
@@ -375,210 +288,5 @@ class AnthropicOnFoundryChatModelTest {
               ConnectorException connectorEx = (ConnectorException) thrown;
               assertThat(connectorEx.getErrorCode()).isEqualTo("TRANSPORT_ERROR");
             });
-  }
-
-  // -------------------------------------------------------------------------
-  // MessageCreateParams request-side translation tests
-  // -------------------------------------------------------------------------
-
-  @Test
-  void buildParams_passes_system_message_to_anthropic_system_field() {
-    // given — adapter passes system text as a plain String to builder.system(String)
-    var message =
-        message(
-            "msg_sys",
-            List.of(ContentBlock.ofText(textBlock("hi"))),
-            StopReason.END_TURN,
-            usage(1, 1));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder()
-            .messages(SystemMessage.from("You are a careful assistant."), UserMessage.from("hi"))
-            .build();
-
-    // when
-    adapter.chat(request);
-
-    // then
-    verify(mockMessageService).create(paramsCaptor.capture());
-    MessageCreateParams params = paramsCaptor.getValue();
-
-    // The adapter calls builder.system(String), so the System sealed type is in isString() variant
-    assertThat(params.system()).isPresent();
-    MessageCreateParams.System system = params.system().get();
-    assertThat(system.isString()).isTrue();
-    assertThat(system.asString()).isEqualTo("You are a careful assistant.");
-  }
-
-  @Test
-  void buildParams_groups_tool_result_messages_into_single_user_message() {
-    // given — two consecutive tool-result messages must be merged into one user-role MessageParam
-    // containing two ToolResultBlockParam content blocks (Anthropic API requirement)
-    var message =
-        message(
-            "msg_tr",
-            List.of(ContentBlock.ofText(textBlock("ok"))),
-            StopReason.END_TURN,
-            usage(1, 1));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder()
-            .messages(
-                UserMessage.from("Get my order status"),
-                AiMessage.builder()
-                    .toolExecutionRequests(
-                        List.of(
-                            ToolExecutionRequest.builder()
-                                .id("call_1")
-                                .name("get_order")
-                                .arguments("{}")
-                                .build(),
-                            ToolExecutionRequest.builder()
-                                .id("call_2")
-                                .name("get_user")
-                                .arguments("{}")
-                                .build()))
-                    .build(),
-                ToolExecutionResultMessage.from("call_1", "get_order", "{\"status\":\"shipped\"}"),
-                ToolExecutionResultMessage.from("call_2", "get_user", "{\"name\":\"Alice\"}"))
-            .build();
-
-    // when
-    adapter.chat(request);
-
-    // then
-    verify(mockMessageService).create(paramsCaptor.capture());
-    MessageCreateParams params = paramsCaptor.getValue();
-
-    List<MessageParam> messages = params.messages();
-    // Expected order: user("Get my order status"), assistant(tool-use blocks), user(merged tool
-    // results)
-    assertThat(messages).hasSize(3);
-
-    MessageParam mergedToolResults = messages.get(2);
-    assertThat(mergedToolResults.role()).isEqualTo(MessageParam.Role.USER);
-    assertThat(mergedToolResults.content().isBlockParams()).isTrue();
-
-    List<ContentBlockParam> blocks = mergedToolResults.content().asBlockParams();
-    assertThat(blocks).hasSize(2);
-
-    // First block: tool result for call_1
-    assertThat(blocks.get(0).isToolResult()).isTrue();
-    ToolResultBlockParam result1 = blocks.get(0).asToolResult();
-    assertThat(result1.toolUseId()).isEqualTo("call_1");
-
-    // Second block: tool result for call_2
-    assertThat(blocks.get(1).isToolResult()).isTrue();
-    ToolResultBlockParam result2 = blocks.get(1).asToolResult();
-    assertThat(result2.toolUseId()).isEqualTo("call_2");
-  }
-
-  @Test
-  void buildParams_translates_assistant_tool_calls_into_tool_use_blocks() {
-    // given — an AiMessage with a tool execution request must produce an assistant-role
-    // MessageParam
-    // whose content contains a ToolUseBlockParam with the right id, name, and parsed input
-    var message =
-        message(
-            "msg_tu",
-            List.of(ContentBlock.ofText(textBlock("ok"))),
-            StopReason.END_TURN,
-            usage(1, 1));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder()
-            .messages(
-                UserMessage.from("Get the date"),
-                AiMessage.builder()
-                    .toolExecutionRequests(
-                        List.of(
-                            ToolExecutionRequest.builder()
-                                .id("toolu_01")
-                                .name("GetDateAndTime")
-                                .arguments("{\"timezone\":\"UTC\"}")
-                                .build()))
-                    .build())
-            .build();
-
-    // when
-    adapter.chat(request);
-
-    // then
-    verify(mockMessageService).create(paramsCaptor.capture());
-    MessageCreateParams params = paramsCaptor.getValue();
-
-    List<MessageParam> messages = params.messages();
-    // Expected: user message, assistant message
-    assertThat(messages).hasSize(2);
-
-    MessageParam assistantMsg = messages.get(1);
-    assertThat(assistantMsg.role()).isEqualTo(MessageParam.Role.ASSISTANT);
-    assertThat(assistantMsg.content().isBlockParams()).isTrue();
-
-    List<ContentBlockParam> blocks = assistantMsg.content().asBlockParams();
-    // Expect exactly one ToolUseBlockParam (no text prefix since AiMessage has no text)
-    assertThat(blocks).hasSize(1);
-    assertThat(blocks.get(0).isToolUse()).isTrue();
-
-    ToolUseBlockParam toolUse = blocks.get(0).asToolUse();
-    assertThat(toolUse.id()).isEqualTo("toolu_01");
-    assertThat(toolUse.name()).isEqualTo("GetDateAndTime");
-
-    // Input is stored as additional properties on ToolUseBlockParam.Input
-    assertThat(toolUse.input()._additionalProperties()).containsKey("timezone");
-    assertThat(toolUse.input()._additionalProperties().get("timezone").toString()).contains("UTC");
-  }
-
-  @Test
-  void buildParams_default_max_tokens_when_neither_config_nor_request_sets_one() {
-    // given — model with null parameters → should fall back to 1024
-    var message =
-        message(
-            "msg_def",
-            List.of(ContentBlock.ofText(textBlock("ok"))),
-            StopReason.END_TURN,
-            usage(1, 1));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var adapterNoParams =
-        new AnthropicOnFoundryChatModel(mockClient, new AnthropicModel("claude-sonnet-4-6", null));
-
-    var request = ChatRequest.builder().messages(UserMessage.from("hi")).build();
-
-    // when
-    adapterNoParams.chat(request);
-
-    // then
-    verify(mockMessageService).create(paramsCaptor.capture());
-    MessageCreateParams params = paramsCaptor.getValue();
-
-    assertThat(params.maxTokens()).isEqualTo(1024L);
-  }
-
-  @Test
-  void buildParams_chat_request_max_tokens_overrides_model_config() {
-    // given — setUp() wires modelConfig with maxTokens=1024; ChatRequest overrides to 4096
-    var message =
-        message(
-            "msg_ov",
-            List.of(ContentBlock.ofText(textBlock("ok"))),
-            StopReason.END_TURN,
-            usage(1, 1));
-    when(mockMessageService.create(any(MessageCreateParams.class))).thenReturn(message);
-
-    var request =
-        ChatRequest.builder().messages(UserMessage.from("hi")).maxOutputTokens(4096).build();
-
-    // when
-    adapter.chat(request);
-
-    // then
-    verify(mockMessageService).create(paramsCaptor.capture());
-    MessageCreateParams params = paramsCaptor.getValue();
-
-    assertThat(params.maxTokens()).isEqualTo(4096L);
   }
 }
